Revert file name quoting
The file doesn't have to be quoted.
First of all, it doesn't work. terraform process believes the file name
includes a `'`, which is incorrect.
Second, there is no shell injection vulnerability here because
`terraform` isn't a shell. It doesn't execute what's passed to it as
arguments.

This "optimization" wasn't needed.