Fix the if condition
Found this in docs:

```
You need to use specific syntax to tell GitHub to evaluate an expression
rather than treat it as a string.

${{ <expression> }}

When you use expressions in an if conditional, you may omit the ${{ }}
expression syntax because GitHub Actions automatically evaluates the if
conditional as an expression. Using the ${{ }} expression syntax turns
the contents into a string, and strings are truthy. For example, if:
true && ${{ false }} will evaluate to true.
```

Not sure why it evaluates `env.PUBLISH == 'yes'` but doesn't evaluate
`env.PUBLISH == 'yes' && ${{ matrix.codename }} == 'jammy'`.