Publish only a summary of changes to GitHub (#72)
* Skip development test

* Bump version: 2.9.7 → 2.9.8

* Publish only a summary of changes to GitHub

`terraform plan` is versbose and the output is useful.
But normally, a user is interested in the expected changes only.
When ih-plan publishes full stdout to a GitHub pull request
comment,often it's too big and has to go ti gist.
This is not optimal UX, so instead of the full output, ih-plan should
publish a summary of changes only.

The full output is still available in GitHub Actions.

* Bump version: 2.9.8 → 2.10.0

* Make existing tests pass