Add scripts for building infrahouse-toolkit on arm64
Well, I couldn't install docker on a macos-14 GitHub Actions runner.
But I still need infrahouse-toolkit packages on arm64.
As a workaround, I created a `build-arm64.sh` build script. It's supposed to run on a MacOS laptop with arm64 chip.
The script builds the package for Ubuntu focal, jammy, and noble and uploads the package to https://release-${OS_VERSION}.infrahouse.com/